requirements.txt
bcrypt==4.0.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT==0.3.2
Flask-Login==0.6.2
Flask-WTF==1.2.1
fonttools==4.43.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.3
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.0
WTForms==3.0.1
